app/processors/office.py
import os
from docx import Document
from openpyxl import load_workbook
from pptx import Presentation
from app.processors.base import BaseProcessor
from app.models.responses import ProcessResult, MetadataInfo

class OfficeProcessor(BaseProcessor):
    """Processor for Office documents (Word, Excel, PowerPoint)."""

    def remove_metadata(self, file_path: str) -> ProcessResult:
        ext = os.path.splitext(file_path)[1].lower()
        if ext == '.docx':
            return self._process_docx(file_path)
        elif ext == '.xlsx':
            return self._process_xlsx(file_path)
        elif ext == '.pptx':
            return self._process_pptx(file_path)
        else:
            raise ValueError(f"Unsupported office file type: {ext}")

    def _process_docx(self, file_path):
        doc = Document(file_path)
        cp = doc.core_properties
        cp.author = None
        cp.last_modified_by = None
        cp.comments = None
        # ... and other properties

        processed_file_path = file_path.replace(os.path.basename(file_path), "processed_" + os.path.basename(file_path))
        doc.save(processed_file_path)
        # ... create ProcessResult
        pass

    def _process_xlsx(self, file_path):
        wb = load_workbook(file_path)
        cp = wb.properties
        cp.creator = None
        cp.lastModifiedBy = None
        # ... and other properties

        processed_file_path = file_path.replace(os.path.basename(file_path), "processed_" + os.path.basename(file_path))
        wb.save(processed_file_path)
        # ... create ProcessResult
        pass

    def _process_pptx(self, file_path):
        prs = Presentation(file_path)
        cp = prs.core_properties
        cp.author = None
        cp.last_modified_by = None
        # ... and other properties

        processed_file_path = file_path.replace(os.path.basename(file_path), "processed_" + os.path.basename(file_path))
        prs.save(processed_file_path)
        # ... create ProcessResult
        pass


    def get_metadata_info(self, file_path: str) -> MetadataInfo:
        # Implementation for metadata info extraction
        pass

    def validate_file(self, file_path: str) -> bool:
        # Implementation for file validation
        pass
